% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ames.R
\docType{data}
\name{ames}
\alias{ames}
\title{The Ames Iowa Housing data from the AmesHousing package}
\format{
A dataframe with 2930 rows and 81 variables:
}
\source{
\url{http://jse.amstat.org/v19n3/decock/DataDocumentation.txt}
}
\usage{
ames
}
\description{
The Ames Iowa Housing data from the AmesHousing package
}
\keyword{datasets}
