% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer}
\alias{customer}
\title{Customer data}
\format{A data frame with 5000 rows and 59 variables.}
\usage{
customer
}
\description{
An artificial dataset containing 59 variables for 5000 customers.
}
\keyword{datasets}
