% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\docType{data}
\name{stops}
\alias{stops}
\title{Data on traffic stops by police}
\format{
A tibble with 211211 rows and 15 variables:
\describe{
\item{Check_website}{dbl openpolicing.stanford.edu}
}
}
\source{
\url{https://openpolicing.stanford.edu}
}
\usage{
stops
}
\description{
These data are a small subset extracted from the openpolicing.stanford.edu dataset and contain traffic stops by police aggregated for each county in Mississippi.
}
\keyword{datasets}
