% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBootcamp-package.R
\docType{package}
\name{RBootcamp-package}
\alias{RBootcamp}
\alias{RBootcamp-package}
\title{RBootcamp: MBAN R Bootcamp Package}
\description{
Data, Tutorials, and other materials.
}
\author{
\strong{Maintainer}: Yigit Aydede \email{yigit.aydede@smu.ca}

}
\keyword{internal}
